% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{d_mallows}
\alias{d_mallows}
\title{This function estimate the distance between
two object of kind symbolic_histogram, of package
RSDA}
\usage{
d_mallows(x, y)
}
\arguments{
\item{x}{a symbolic_histogram object}

\item{y}{a symbolic_histogram object}
}
\value{
return the Mallows Distance between the objects
}
\description{
This function estimate the distance between
two object of kind symbolic_histogram, of package
RSDA
}
