% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{mean.symbolic_histogram}
\alias{mean.symbolic_histogram}
\title{Function to estimate the mean of symbolic_histograms}
\usage{
\method{mean}{symbolic_histogram}(x)
}
\arguments{
\item{x}{vector of symbolic_histogram objects}
}
\value{
a symbolic_histogram objects that is the Frechet mean by Mallows
}
\description{
Function to estimate the mean of symbolic_histograms
}
