% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{median.symbolic_histogram}
\alias{median.symbolic_histogram}
\title{Function to estimate the median of symbolic_histograms}
\usage{
\method{median}{symbolic_histogram}(x)
}
\arguments{
\item{x}{symbolic_histogram objects}
}
\value{
the meadian of the symbolic_histogram
}
\description{
Function to estimate the median of symbolic_histograms
}
