% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{quantile_symbolic_histogram}
\alias{quantile_symbolic_histogram}
\title{Quantile function, it has as purpose to estimate the
quantile given an symbolic_histogram}
\usage{
quantile_symbolic_histogram(x, prob = 0.1)
}
\arguments{
\item{x}{it is a symbolic_histogram object}

\item{prob}{it is the probability of the quantile to estimate}
}
\value{
return a numeric value
}
\description{
Quantile function, it has as purpose to estimate the
quantile given an symbolic_histogram
}
