% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{uniform_histogram}
\alias{uniform_histogram}
\title{Function to generate uniform histograms}
\usage{
uniform_histogram(x, cum_w = NULL, length_quantiles = 0.1)
}
\arguments{
\item{x}{symbolic_histogram object that is going to be uniformed}

\item{cum_w}{cummulative weight that is considered to uniform histogram. It can be NULL, but is used length_quantile in this case}

\item{length_quantiles}{distance beteween the weights to uniform histogram. It can be NULL, but is used cum_w in this case}
}
\value{
symbolic histogram uniformed
}
\description{
Function to generate uniform histograms
}
\keyword{internal}
